Rename BYOF runner to run_byof_repo.py with compat shim
Generic entrypoint for OSS repo containerize + rl-train/datagen workloads.
Keep run_isaac_lab_byof_repo.py as a thin delegate for existing callers.
