+++ npa/src/npa/cli/agent_chat.py
@@ -719,7 +719,7 @@ def format_onboard_solution() -> str:
             "- Step 3 (**live infra**): submit a SkyPilot smoke workflow in tmux with retry gates.",
             "- Generic onboarding command (replace placeholders):",
             "```bash",
-            "npa/.venv/bin/python npa/scripts/run_isaac_lab_byof_repo.py \\",
+            "npa/.venv/bin/python npa/scripts/run_byof_repo.py \\",
             "  --repo-url <repo-url> \\",
             "  --repo-ref <repo-ref> \\",
             "  --registry " + registry + " \\",
@@ -732,11 +732,11 @@ def format_onboard_solution() -> str:
             "```bash",
             "SESSION=solution-live-$(date -u +%Y%m%dT%H%M%SZ)",
             "tmux new -d -s \"$SESSION\"",
-            "tmux send-keys -t \"$SESSION:0.0\" 'set -euo pipefail; sky check; sky gpus list; ATTEMPT=1; while [ $ATTEMPT -le 3 ]; do npa/.venv/bin/python npa/scripts/run_isaac_lab_byof_repo.py --repo-url <repo-url> --repo-ref <repo-ref> --yaml <resource-profile.yaml> --task <task> --iterations 1 --cleanup && break; ATTEMPT=$((ATTEMPT+1)); sleep $((ATTEMPT*20)); done' C-m",
+            "tmux send-keys -t \"$SESSION:0.0\" 'set -euo pipefail; sky check; sky gpus list; ATTEMPT=1; while [ $ATTEMPT -le 3 ]; do npa/.venv/bin/python npa/scripts/run_byof_repo.py --repo-url <repo-url> --repo-ref <repo-ref> --yaml <resource-profile.yaml> --task <task> --iterations 1 --cleanup && break; ATTEMPT=$((ATTEMPT+1)); sleep $((ATTEMPT*20)); done' C-m",
             "```",
             "- LeIsaac validation (datagen): scripted state-machine demos at scale:",
             "```bash",
-            "npa/.venv/bin/python npa/scripts/run_isaac_lab_byof_repo.py \\",
+            "npa/.venv/bin/python npa/scripts/run_byof_repo.py \\",
             "  --repo-url https://github.com/LightwheelAI/leisaac.git \\",
             "  --repo-ref main \\",
             "  --registry " + registry + " \\",
