+++ npa/tests/cli/test_agent.py
@@ -75,7 +75,7 @@ def test_bootstrap_embeds_chat_endpoint() -> None:
     assert "markdownLiteHtml" in source
     assert "Secure basic-auth session" in source
     assert "sparkle" in source
-    assert "run_isaac_lab_byof_repo.py" in source
+    assert "run_byof_repo.py" in source
     assert "For BYOF solution onboarding" in source
     assert "Always use real registry-qualified images" in source
     assert "`<your-registry-id>` placeholders" in source
